Safari-611.1.10/LayoutTests/editing/selection/ios/show-selection-in-transformed-container-2-expected.txt


This test verifies that after focusing a visible input field in a display: flex container that has been translated horizontally out of view, the caret is still visible. To run the test manually, tap the input field and check that the caret shows up.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caretRect.left is 12
PASS caretRect.top is 6
PASS caretRect.width is 2
PASS caretRect.height is 23
PASS successfullyParsed is true

TEST COMPLETE
